A_start donne with heap, Make clean modified to clean *.o and *.d

+++ list.h
@@ -7,7 +7,7 @@ typedef struct List List;
 struct List
 {
     int val;
-    int cost;
+    double cost;
     List *next;
 };
 
@@ -15,9 +15,9 @@ List *initlist();
 
 void append(struct List *list, int nv);
 
-void update(struct List* list, int val, int cost);
+void update(struct List* list, int val, double cost);
 
-void h_pop(struct List* list, int* val, int* cost);
+void h_pop(struct List* list, int* val, double* cost);
 
 int pop(struct List *liste);
